Discordbot/requirements.txt
py-cord
google_api_python_client
protobuf
python-dotenv
wikipedia
yt_dlp